[arcane,accelerator] Move 'AcceleratorMemoryAllocatorBase' in its own file.
